# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/queue/create-service-order-and-enqueue": {
      "post": {
        "operationId": "QueueController_createServiceOrderAndEnqueue",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "创建服务订单并入队（先服务后付）",
        "tags": [
          "queue"
        ]
      }
    },
    "/queue/{id}/set-current": {
      "post": {
        "operationId": "QueueController_setCurrent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "设置当前执行任务索引",
        "tags": [
          "queue"
        ]
      }
    },
    "/queue/{id}/finish-task": {
      "post": {
        "operationId": "QueueController_finishTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "完成当前任务节点",
        "tags": [
          "queue"
        ]
      }
    },
    "/queue/{id}/confirm-complete": {
      "post": {
        "operationId": "QueueController_confirmComplete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "确认整单已完成",
        "tags": [
          "queue"
        ]
      }
    },
    "/queue/{id}/start-first": {
      "post": {
        "operationId": "QueueController_startFirst",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "开始第一步任务",
        "tags": [
          "queue"
        ]
      }
    },
    "/queue/{id}": {
      "delete": {
        "operationId": "QueueController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "移出队列/取消",
        "tags": [
          "queue"
        ]
      }
    },
...
    "/store/inventory/adjust": {
      "post": {
        "operationId": "StoreInventoryController_adjust",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreInventoryAdjustDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StoreInventory"
        ]
      }
    },
    "/store/inventory/logs": {
      "get": {
        "operationId": "StoreInventoryController_logs",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skuId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StoreInventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wash Club API",
    "description": "API for 一体化洗车门店管理平台",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StoreInventoryAdjustDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "description": "商品ID",
            "example": 1
          },
          "skuId": {
            "type": "object",
            "description": "SKU ID（多规格时传）",
            "example": 10,
            "nullable": true
          },
          "change": {
            "type": "number",
            "description": "变更数量（入库/出库为正数，调整可正可负）",
            "example": 5
          },
          "reason": {
            "type": "string",
            "description": "原因",
            "example": "INBOUND"
          },
          "remark": {
            "type": "object",
            "description": "备注",
            "example": "盘点修正",
            "nullable": true
          }
        },
        "required": [
          "productId",
          "change",
          "reason"
        ]
      },
...
    }
  }
}